Removed hello page, made index page more useful as a test
