+++ src/pages/index.tsx
@@ -1,68 +1,35 @@
+import React from "react"
 import type { NextPage } from "next"
-import Image from "next/image"
 import Link from "next/link"
+import { Button, GovBanner } from "@trussworks/react-uswds"
+
+import { useHello } from "hooks/hello"
 import styles from "styles/home.module.scss"
 
-const Home: NextPage = () => {
+const isError = (error: unknown): error is Error => error instanceof Error
+
+const HelloPage: NextPage = () => {
+  const { data, error } = useHello()
+
   return (
     <div className={styles.container}>
       <main className={styles.main}>
-        <h1 className={styles.title}>Welcome to the Next Front-End</h1>
+        <h1 className={styles.title}>Hello, welcome to the JEDI Project</h1>
 
         <p className={styles.description}>
-          Get started by editing{" "}
-          <code className={styles.code}>pages/index.tsx</code>
+          Greetings from the Backend:{" "}
+          <code className={styles.code}>
+            {data ? data.hello : isError(error) ? error.message : "loading..."}
+          </code>
         </p>
 
-        <Link href="/hello">
-          <a className={styles.card}>Test Page</a>
-        </Link>
-
-        <div className={styles.grid}>
-          <a href="https://nextjs.org/docs" className={styles.card}>
-            <h2>Documentation &rarr;</h2>
-            <p>Find in-depth information about Next.js features and API.</p>
-          </a>
-
-          <a href="https://nextjs.org/learn" className={styles.card}>
-            <h2>Learn &rarr;</h2>
-            <p>Learn about Next.js in an interactive course with quizzes!</p>
-          </a>
-
-          <a
-            href="https://github.com/vercel/next.js/tree/master/examples"
-            className={styles.card}
-          >
-            <h2>Examples &rarr;</h2>
-            <p>Discover and deploy boilerplate example Next.js projects.</p>
-          </a>
-
-          <a
-            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-          >
-            <h2>Deploy &rarr;</h2>
-            <p>
-              Instantly deploy your Next.js site to a public URL with Vercel.
-            </p>
-          </a>
-        </div>
+        <a href="/incidents" className={styles.card}>
+            <h2>Incidents &rarr;</h2>
+            <p>All incident entries</p>
+        </a>        
       </main>
-
-      <footer className={styles.footer}>
-        <a
-          href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Powered by{" "}
-          <span className={styles.logo}>
-            <Image src="/vercel.svg" alt="Vercel Logo" width={72} height={16} />
-          </span>
-        </a>
-      </footer>
     </div>
   )
 }
 
-export default Home
+export default HelloPage
